refactor: 활동 내역 관련 이벤트 클래스 디렉토리 변경
