+++ src/main/java/com/codeit/monew/domain/useractivity/event/ArticleViewedEvent.java
@@ -1,4 +1,4 @@
-package com.codeit.monew.global.event;
+package com.codeit.monew.domain.useractivity.event;
 
 import com.codeit.monew.domain.article.ArticleSource;
 import java.time.Instant;
