+++ src/main/java/com/codeit/monew/domain/useractivity/event/CommentLikedEvent.java
@@ -1,4 +1,4 @@
-package com.codeit.monew.global.event;
+package com.codeit.monew.domain.useractivity.event;
 
 import java.time.Instant;
 import java.util.UUID;
