+++ src/main/java/com/codeit/monew/domain/useractivity/listener/UserActivityEventListener.java
@@ -7,14 +7,14 @@ import com.codeit.monew.domain.useractivity.entity.UserActivity.CommentLikeInfo;
 import com.codeit.monew.domain.useractivity.entity.UserActivity.SubscriptionInfo;
 import com.codeit.monew.domain.useractivity.mapper.UserActivityMapper;
 import com.codeit.monew.domain.useractivity.repository.UserActivityRepository;
-import com.codeit.monew.global.event.ArticleViewedEvent;
-import com.codeit.monew.global.event.CommentCreatedEvent;
-import com.codeit.monew.global.event.CommentLikedCancelEvent;
-import com.codeit.monew.global.event.CommentLikedEvent;
-import com.codeit.monew.global.event.CommentUpdatedEvent;
-import com.codeit.monew.global.event.InterestSubscribedEvent;
-import com.codeit.monew.global.event.InterestUnSubscribedEvent;
-import com.codeit.monew.global.event.UserRegisteredEvent;
+import com.codeit.monew.domain.useractivity.event.ArticleViewedEvent;
+import com.codeit.monew.domain.useractivity.event.CommentCreatedEvent;
+import com.codeit.monew.domain.useractivity.event.CommentLikedCancelEvent;
+import com.codeit.monew.domain.useractivity.event.CommentLikedEvent;
+import com.codeit.monew.domain.useractivity.event.CommentUpdatedEvent;
+import com.codeit.monew.domain.useractivity.event.InterestSubscribedEvent;
+import com.codeit.monew.domain.useractivity.event.InterestUnSubscribedEvent;
+import com.codeit.monew.domain.useractivity.event.UserRegisteredEvent;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.Document;
