+++ src/test/java/com/codeit/monew/domain/article/service/ArticleServiceTest.java
@@ -27,8 +27,7 @@ import com.codeit.monew.domain.interest.entity.Interest;
 import com.codeit.monew.domain.interest.repository.InterestRepository;
 import com.codeit.monew.domain.user.entity.User;
 import com.codeit.monew.domain.user.repository.UserRepository;
-import com.codeit.monew.global.event.ArticleViewedEvent;
-import com.codeit.monew.global.event.UserRegisteredEvent;
+import com.codeit.monew.domain.useractivity.event.ArticleViewedEvent;
 import com.codeit.monew.global.exception.article.ArticleNotFoundException;
 import com.codeit.monew.global.exception.Interest.InterestNotFoundException;
 import com.codeit.monew.global.exception.user.UserNotFoundException;
