+++ src/test/java/com/codeit/monew/domain/comment/service/CommentLikeServiceTest.java
@@ -16,9 +16,8 @@ import com.codeit.monew.domain.comment.repository.CommentLikeRepository;
 import com.codeit.monew.domain.comment.repository.CommentRepository;
 import com.codeit.monew.domain.user.entity.User;
 import com.codeit.monew.domain.user.repository.UserRepository;
-import com.codeit.monew.global.event.ArticleViewedEvent;
-import com.codeit.monew.global.event.CommentLikedCancelEvent;
-import com.codeit.monew.global.event.CommentLikedEvent;
+import com.codeit.monew.domain.useractivity.event.CommentLikedCancelEvent;
+import com.codeit.monew.domain.useractivity.event.CommentLikedEvent;
 import com.codeit.monew.global.exception.comment.CommentLikeAlreadyExistsException;
 import com.codeit.monew.global.exception.comment.CommentLikeNotFoundException;
 import com.codeit.monew.global.exception.comment.CommentNotFoundException;
