+++ src/test/java/com/codeit/monew/domain/user/service/UserServiceTest.java
@@ -13,11 +13,10 @@ import com.codeit.monew.domain.user.dto.UserUpdateRequest;
 import com.codeit.monew.domain.user.entity.User;
 import com.codeit.monew.domain.user.mapper.UserMapper;
 import com.codeit.monew.domain.user.repository.UserRepository;
-import com.codeit.monew.global.event.UserRegisteredEvent;
+import com.codeit.monew.domain.useractivity.event.UserRegisteredEvent;
 import com.codeit.monew.global.exception.ErrorCode;
 import com.codeit.monew.global.exception.user.DuplicateEmailException;
 import com.codeit.monew.global.exception.user.PasswordMismatchException;
-import com.codeit.monew.global.exception.user.UserAccessDeniedException;
 import com.codeit.monew.global.exception.user.UserNotFoundException;
 import java.time.Instant;
 import java.util.Optional;
@@ -26,7 +25,6 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
